+++ Rhombus/src/Rhombus/Debug/Instrumentor.h
@@ -165,7 +165,7 @@ namespace rhombus
 			auto start = FloatingPointMicroseconds{ m_StartTimepoint.time_since_epoch() };
 			auto elapsedTime = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch() - std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch();
 
-			uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+			size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
 			Instrumentor::Get().WriteProfile({ m_Name, start, elapsedTime, std::this_thread::get_id() });
 
 			m_Stopped = true;
